Add trace-level sampling
When using high sampling rates, beeline is very inneficient. Span
objects, data is collected, just to be send to the tracer and
discarded.

Also a lot of the sampling is based on the trace id, so all spans
in a single trace all recorded or none. In such cases, doing
so much work when we know ahead of time that the span is going
to be discarded is a waste.

This adds sampling at the trace level. When a trace stats, gets
a random trace id, and if a trace sampling hook is set, it is
invoked to determine if the trace should capture or not.

If it is determined to not be captured, all other beeline
calls on the trace become a no-op, no spans are created and
no context is recorded.